requirements.txt
aiohappyeyeballs==2.4.3
aiohttp==3.10.10
aiosignal==1.3.1
anyio==4.8.0
asttokens==3.0.0
attrs==24.2.0
bcrypt==4.2.1
beautifulsoup4==4.12.3
bs4==0.0.2
certifi==2024.7.4
cffi==1.17.1
charset-normalizer==3.3.2
colorama==0.4.6
comm==0.2.2
cryptography==44.0.0
debugpy==1.8.12
decorator==5.1.1
defusedxml==0.7.1
dig==0.1.10
dill==0.3.8
docker==7.1.0
docopt==0.6.2
executing==2.2.0
frozenlist==1.5.0
futures==3.0.5
geographiclib==2.0
geoip2==4.8.0
geopy==2.4.1
h11==0.14.0
httpcore==1.0.7
httpie==3.2.4
httpx==0.28.1
idna==3.7
ipykernel==6.29.5
ipython==8.31.0
jedi==0.19.2
joblib==1.4.2
jupyter_client==8.6.3
jupyter_core==5.7.2
loguru==0.7.2
lxml==5.3.0
markdown-it-py==3.0.0
matplotlib-inline==0.1.7
maxminddb==2.6.2
mdurl==0.1.2
multidict==6.1.0
multiprocess==0.70.16
nest-asyncio==1.6.0
netcat==0.0.1a0
numpy==2.1.3
outcome==1.3.0.post0
packaging==24.2
pandas==2.2.3
paramiko==3.5.1
parso==0.8.4
pathos==0.3.2
pillow==11.0.0
platformdirs==4.3.6
pox==0.3.4
ppft==1.7.6.8
prompt_toolkit==3.0.50
propcache==0.2.0
psutil==6.1.1
pure_eval==0.2.3
pycparser==2.22
PyExecJS==1.5.1
Pygments==2.18.0
PyNaCl==1.5.0
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-telegram-bot==21.10
pytz==2024.2
pywin32==306
pyzmq==26.2.0
requests==2.32.3
requests-toolbelt==1.0.0
rich==13.9.4
scapy==2.6.0
scikit-learn==1.6.1
scipy==1.15.1
selenium==4.28.1
setuptools==75.2.0
six==1.16.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.6
sshuttle==1.1.2
stack-data==0.6.3
threadpoolctl==3.5.0
tornado==6.4.2
traitlets==5.14.3
translate-api==4.9.5
trio==0.28.0
trio-websocket==0.11.1
typing_extensions==4.12.2
tzdata==2025.1
urllib3==2.2.2
wcwidth==0.2.13
webdriver-manager==4.0.2
websocket-client==1.8.0
wget==3.2
win32-setctime==1.1.0
wsproto==1.2.0
yarl==1.16.0
